refactor(#555): extract preset filename helpers to application::preset_file
The filename ↔ preset-name helpers (`PRESET_EXTENSION`,
`sanitize_for_filename`, `preset_filename`, `preset_save_path`) used
to live in `adapter-gui::chain_preset_wiring`, right next to the UI
callback that called them. That placement kept the dispatcher from
resolving a preset's on-disk path on its own, so the GUI was the only
caller that could do `fs::write` / `fs::remove_file` — a violation of
the "backend transport-agnostic" + "tela sem regra de negócio"
laws (#436, #555 audit).

This commit lifts the helpers into a new `application::preset_file`
module so the dispatcher and the GUI share one source of truth.

What changes:
* New `crates/application/src/preset_file.rs` with the four
  helpers, plus dedicated unit tests
  (`preset_file_tests.rs`, 4 passing) pinning the
  filesystem-illegal char set, the .yaml extension and the
  user-name-preserving sanitisation rule (issue #510).
* `adapter-gui/src/chain_preset_wiring.rs` removes the local copies
  and `pub(crate) use application::preset_file::{preset_filename,
  preset_save_path}` so existing in-crate callers
  (`preset_save_wiring`, the test module) keep working unchanged.

Pure refactor; no behaviour change. Sets the stage for the next
commits, which move the actual `fs::remove_file` / `fs::write` calls
from the GUI into the `SaveChainPreset` / `DeleteChainPreset`
dispatcher handlers.
